Remove apis and related functions for refactoring
