+++ SUI/suiapi.py
@@ -1,6 +1,7 @@
 import requests
 from Walrus import walrus
-from DB import models, schemas
+# from DB import models, schemas
+from DB import models, schemas_legacy
 
 BASE_URL = "http://localhost:8080"
 RAGCOON_STAGE_ID = "0x68e7482eb88d2bfe57481a8078ed447bc50c00f7487d9484bc00b9e49c0c7986"
@@ -41,7 +42,7 @@ def add_ai(ai_id: str, creator_address: str):
     return digest
 
 
-def add_blob(ai: schemas.AITableCreate ,ai_id : str, embed: str):
+def add_blob(ai: schemas_legacy.AITableCreate ,ai_id : str, embed: str):
     url = BASE_URL + "/movecall/add_blob_id"  # The URL of the REST API you want to call
     res = walrus.send_data(str(embed))
